Nest Google credentials in SM under a top-level JSON key
The pipeline around the external secrets Kubernetes operator tries to be
overly "helpful" by treating top-level JSON as individual secrets, when
what we _really_ want is the whole JSON as a single entity.

This nests the Google credentials JSON under a top-level key, which we
can then extract using the external secrets operator as a single file.