Datastore schema: disallow additional properties
This updates the schema to reflect that no additional properties should
be present in any JSON documents above and beyond those explicitly
specified in the schema.

Also updates Github lint workflow to always run even when no relevant
files have been changed (as this is now a required workflow for branch
protection purposes).