Add initial set of boosts to Discovery Engine
These are used to promote or demote documents based on their metadata
where we know that certain kinds of documents contain less
important/"mainstream" information and get in the way of users finding
more relevant content.

This particular configuration is predominantly adapted from one that
provided good results during out proof-of-concept phase, which in turn
is quite similar to the boosts originally found in Search API's
`config/query/boosting.yml`.