Migrate to `google_discovery_engine_data_store` resource
We are currently semi-manually creating all Discovery Engine (Google
Vertex AI Search) resources using the REST API provider as when we
started work on this project, the first party Google provider didn't
offer any of the resources we needed yet. This is (slowly) starting to
change, so we want to migrate to the first party resources.

- Update Google Terraform provider to latest available version
- Add `google_discovery_engine_data_store` resource for the datastore,
  plus an import block to capture the existing resource
- Remove `restapi_object.discovery_engine_datastore` resource from
  `google_discovery_engine_restapi` module, and add removed block with
  lifecycle rule to ensure it doesn't get deleted during the swap out
- Remove outputs from `google_discovery_engine_restapi` module (these
  were just making up hardcoded names anyway based on convention, now
  that the datastore lives in the main `environment` module we can just
  do that in locals there)