Add `public_timestamp_datetime` field to schema
We have been using manual boosting and sorting on the `public_timestamp`
field as an integer until now.

Discovery Engine now supports datetime fields in RFC3339 format for both
a new freshness boosting feature as well as sorting, so we should be
able to deprecate the existing `public_timestamp` field and use this one
going forwards.