Remove JSON schema validation on CI
We originally added this as an extra guard step to "shift left"
prevention of schema errors. However, the validation now causes issues
because Discovery Engine's schema semantics are subtly different from
vanilla JSON Schema. For example, Discovery Engine field types are a
superset of those allowed in JSON Schema, and we want to start using the
`datetime` field type.

This will not cause any major headaches as our schema is quite stable at
this point anyway, and any issues with the schema would still be caught
further right in the CD pipeline anyway (Discovery Engine would refuse
to accept the schema change at the apply stage).