Add service account for UI prototype
We'll soon start working on a UI prototype for some upcoming interface
improvements for search.

This will (probably) live on Heroku, and get its own set of keys to
access GCP and relevant Discovery Engine resources on a read-only basis.

For now, it gets the minimum set of permissions necessary to perform
search and autocomplete requests.