Add separate engine for UI prototype
We want to be able to experiment with some advanced Vertex AI Search
features in the UI prototype. These come at a cost, and we've
specifically disabled them in the production engine to make sure we
don't accidentally get charged.

This creates a new engine with the features enabled (enterprise tier and
LLM add-on) specifically for use by the UI prototype, leaving the
existing production engine unchanged.